Delete the integration test
It was going to become really unwieldy to implement. Ultimately, the
architecture of that level of integration test has to be done the way
that cob spec does it (i.e. running a separate java process and
inspecting the results). That may be desirable as a lighter-weight way
to make sure everything is wired up, but not right now.