Rewrite the whole chain to use InputStreams
The header and the status line are still generated with Stream<String>'s
though the value is a bit more dubious. This is then converted to a
ByteArrayInputStream after being joined with CRLF's.  This all happens
inside the HttpProtocol which is really the place that should know about
the use of CRLF's.