Start adding an HttpInputStream class
The details of how to parse an stream of octets for Http are sort of
complex, since you can't just depend on the buffering.  The request line
and header section need to parsed as text, and the delineation between
the text portion and the body is two CRLFs in a row. Then everything is
binary data and must be returned as bytes. So I'm encapsulating it in
this class.